requirements.txt
et-xmlfile==1.1.0
lxml==4.9.2
openpyxl==3.1.2
Pillow==9.5.0
python-pptx==0.6.21
PyYAML==6.0
XlsxWriter==3.1.1
